# PERF: reindex with MultiIndex