@twoertwein 2022-03-05
I think this PR made pyright unhappy, but I honestly don't see how this PR affected _mixins.py:

> pandas/core/arrays/_mixins.py:185:26 - warning: Instance methods should take a "self" parameter (reportSelfClsParameterName)

If the solution is to ignore this warning, you can add
`# pyright: reportSelfClsParameterName = false`
at the top of the file.