@twoertwein 2022-03-05
> i think was a prior PR

Yes, you are right: #46214 If no one beats me, I will create a PR to rename cls to self